Added a Universal Player Controller to Resolve Multiple Bugs
Added ability to change car and its stats.
Added ability to have multiple players at once (multiplayer not implemented, but the code should now allow for it)

Fixed Transitions
REMOVED POSSESSION ISSUES
-> Moved control for both classes into a single player controller
Flying is now bugged af, but that can be fixed
TODO:
Fix flying
Make it so the plane can't shoot if it isn't active.